feat: refactor PostComments component; extract AddComment for better separation of concerns and improve comment handling logic
