+++ src/components/ui/button.tsx
@@ -24,6 +24,7 @@ const buttonVariants = cva(
         sm: "h-9 rounded-md px-3",
         lg: "h-11 rounded-md px-8",
         icon: "h-10 w-10",
+        nothing: "",
       },
     },
     defaultVariants: {
