+++ src/hooks/useComments.ts
@@ -1,13 +1,11 @@
 import { IComments } from "@/types/posts";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { useMutation } from "@tanstack/react-query";
 import { useState } from "react";
 import useToken from "./useToken";
 
 export default function useComments(comments: IComments[], postId: number) {
-  const queryClient = useQueryClient();
   const commentsContent = comments?.map((comment) => comment.content);
   const [commentsState, setComments] = useState<string[]>(commentsContent);
-  queryClient.setQueryData(["comments", postId], commentsContent);
   const token = useToken();
 
   const mutatedFunc = useMutation({
@@ -32,7 +30,7 @@ export default function useComments(comments: IComments[], postId: number) {
     onMutate: async (data) => {
       setComments((prev) => [...prev, data.newComment]);
     },
-    onError: (error, newComment, context) => {
+    onError: () => {
       setComments((prev) => prev.slice(0, prev.length - 1));
     },
   });
